Development

+++ repository/IngSoft2-Model/back/src/main/java/api/back/GrupoController.java
@@ -50,7 +50,7 @@ public class GrupoController {
         for (String email : miembrosEmails) {
             User usuario = userService.findByEmail(email);
             LocalDate fechaHoy = LocalDate.now();
-            TransaccionesPendientes transaccionPendiente = new TransaccionesPendientes(0.0, usuario, grupo.getNombre(), "Grupo", fechaHoy);
+            TransaccionesPendientes transaccionPendiente = new TransaccionesPendientes(0.0, usuario, grupo.getNombre(), "Grupo", fechaHoy,null,null);
             // Establecer el correo del creador como sentByEmail
             transaccionPendiente.setSentByEmail(creadorEmail);
             transaccionPendiente.setGrupoId(grupo.getId());
@@ -115,9 +115,19 @@ public class GrupoController {
     
         // Extrae los demás datos del JSON
         String motivo = (String) payload.get("motivo");
-        LocalDate fecha = LocalDate.parse((String) payload.get("fecha")); // Asegúrate de que la fecha esté en formato ISO
+        LocalDate fecha = LocalDate.parse((String) payload.get("fecha"));
         String categoria = (String) payload.get("categoria");
         String tipoGasto = (String) payload.get("tipoGasto");
+        String monedaOriginal = (String) payload.get("monedaOriginal");
+        Double montoOriginal;
+        Object montoOriginalObj = payload.get("montoOriginal");
+        if (montoOriginalObj instanceof String) {
+            montoOriginal = Double.parseDouble((String) montoOriginalObj);
+        } else if (montoOriginalObj instanceof Number) {
+            montoOriginal = ((Number) montoOriginalObj).doubleValue();
+        } else {
+            return ResponseEntity.badRequest().body(null); // Valor no válido
+        }
     
         // Convierte el valor de grupo a Long, similar a como hiciste antes
         Long grupoId = ((Number) payload.get("grupo")).longValue(); // Obtén el ID del grupo desde el JSON
@@ -129,7 +139,7 @@ public class GrupoController {
         }
     
         // Crea una nueva transacción grupal
-        GrupoTransacciones grupoTransaccion = new GrupoTransacciones(valor, motivo, fecha, categoria, tipoGasto, usuarioEmail);
+        GrupoTransacciones grupoTransaccion = new GrupoTransacciones(valor, motivo, fecha, categoria, tipoGasto, usuarioEmail,monedaOriginal,montoOriginal);
         grupoTransaccion.setGrupo(grupo); // Establece la relación con el grupo
     
         // Agrega la transacción a la lista de transacciones del grupo
@@ -142,11 +152,16 @@ public class GrupoController {
     }
 
     @GetMapping("/{grupoId}/transacciones")
-    public ResponseEntity<List<GrupoTransacciones>> obtenerTransaccionesPorGrupo(@PathVariable Long grupoId) {
-        // Busca el grupo por su ID
+    public ResponseEntity<List<GrupoTransacciones>> obtenerTransaccionesPorGrupo(@PathVariable Long grupoId, Authentication authentication) {
+        String emailUsuario = authentication.getName();
         Grupo grupo = grupoService.findById(grupoId);
         if (grupo == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Collections.emptyList()); // Retorna 404 si el grupo no existe
+        }
+        boolean perteneceAlGrupo = grupo.getUsuarios().stream()
+        .anyMatch(usuario -> usuario.getEmail().equals(emailUsuario));
+        if (!perteneceAlGrupo) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Collections.emptyList());
         }
         // Obtiene las transacciones asociadas al grupo
         List<GrupoTransacciones> transacciones = grupo.getTransacciones();
@@ -154,28 +169,30 @@ public class GrupoController {
     }
 
     @PostMapping("/{grupoId}/cerrar")
-    public ResponseEntity<String> cerrarGrupo(@PathVariable Long grupoId) {
-        // Buscar el grupo por su ID
+    public ResponseEntity<String> cerrarGrupo(@PathVariable Long grupoId, Authentication authentication) {
+        String emailUsuario = authentication.getName();
         Grupo grupo = grupoService.findById(grupoId);
         if (grupo == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Grupo no encontrado.");
         }
+        boolean perteneceAlGrupo = grupo.getUsuarios().stream()
+            .anyMatch(usuario -> usuario.getEmail().equals(emailUsuario));
 
+        if (!perteneceAlGrupo) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permiso para cerrar este grupo.");
+        }
         List<TransaccionesPendientes> transaccionesPendientes = transaccionesPendientesService.findByGrupoId(grupoId);
         for (TransaccionesPendientes transaccionPendiente : transaccionesPendientes) {
             transaccionesPendientesService.delete(transaccionPendiente.getId());
         }
 
-        // Sumar todas las transacciones del grupo
         double totalTransacciones = grupo.getTransacciones().stream()
         .mapToDouble(GrupoTransacciones::getValor)
         .sum();
 
-        // Calcular el monto a pagar por cada usuario
         int cantidadUsuarios = grupo.getUsuarios().size();
         double montoPorUsuario = totalTransacciones / cantidadUsuarios;
 
-        // Crear una transacción para cada usuario del grupo
         LocalDate fechaHoy = LocalDate.now();
         String categoria = "Gasto Grupal";
         String motivo = grupo.getNombre();
@@ -189,11 +206,12 @@ public class GrupoController {
             nuevaTransaccion.setMotivo(motivo);
             nuevaTransaccion.setTipoGasto(medioDePago);
             nuevaTransaccion.setUser(usuario);
+            nuevaTransaccion.setMonedaOriginal("ARG");
+            nuevaTransaccion.setMontoOriginal(montoPorUsuario);
             
-            // Guardar la transacción
             transaccionesService.createTransaccion(nuevaTransaccion, usuario.getEmail());
         }
-        // Cambiar el estado del grupo a cerrado (false)
+
         grupo.setEstado(false);
         grupoService.save(grupo); 
         
@@ -201,14 +219,17 @@ public class GrupoController {
     }
 
     @DeleteMapping("/transaccion/{transaccionId}")
-    public ResponseEntity<String> eliminarGrupoTransaccion(@PathVariable Long transaccionId) {
-        // Busca la transacción por su ID
+    public ResponseEntity<String> eliminarGrupoTransaccion(@PathVariable Long transaccionId, Authentication authentication) {
+        String emailUsuario = authentication.getName();
         GrupoTransacciones transaccion = grupoTransaccionesService.findById(transaccionId);
+
         if (transaccion == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Transacción no encontrada.");
         }
+        if (!emailUsuario.equals(transaccion.getUsers())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permiso para eliminar esta transacción.");
+        }
 
-        // Elimina la transacción
         grupoTransaccionesService.delete(transaccionId);
         return ResponseEntity.ok("Transacción eliminada exitosamente.");
     }
@@ -217,15 +238,17 @@ public class GrupoController {
     @PutMapping("/transaccion/{transaccionId}")
     public ResponseEntity<GrupoTransacciones> editarGrupoTransaccion(
         @PathVariable Long transaccionId,
-        @RequestBody Map<String, Object> payload
+        @RequestBody Map<String, Object> payload,
+        Authentication authentication
     ) {
-        // Busca la transacción por su ID
+        String emailUsuario = authentication.getName();
         GrupoTransacciones transaccionExistente = grupoTransaccionesService.findById(transaccionId);
         if (transaccionExistente == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null); // Retorna 404 si la transacción no existe
         }
-
-        // Actualiza los valores de la transacción
+        if (!emailUsuario.equals(transaccionExistente.getUsers())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(null);
+        }
         if (payload.containsKey("valor")) {
             Object valorObj = payload.get("valor");
             Double valor;
@@ -233,7 +256,7 @@ public class GrupoController {
                 valor = ((Number) valorObj).doubleValue();
             } else if (valorObj instanceof String) {
                 valor = Double.parseDouble((String) valorObj);
-            } else {
+            } else { 
                 throw new IllegalArgumentException("Tipo de dato no válido para el campo 'valor'");
             }
             transaccionExistente.setValor(valor);
@@ -251,44 +274,60 @@ public class GrupoController {
         if (payload.containsKey("tipoGasto")) {
             transaccionExistente.setTipoGasto((String) payload.get("tipoGasto"));
         }
-
-        // Guarda los cambios
+        if (payload.containsKey("monedaOriginal")) {
+            transaccionExistente.setMonedaOriginal((String) payload.get("monedaOriginal"));
+        }
+        if (payload.containsKey("montoOriginal")) {
+            Object valorObj2 = payload.get("montoOriginal");
+            Double valorOriginal;
+            if (valorObj2 instanceof Number) {
+                valorOriginal = ((Number) valorObj2).doubleValue();
+            } else if (valorObj2 instanceof String) {
+                valorOriginal = Double.parseDouble((String) valorObj2);
+            } else { 
+                throw new IllegalArgumentException("Tipo de dato no válido para el campo 'valor'");
+            }
+            transaccionExistente.setMontoOriginal(valorOriginal);
+        }
         GrupoTransacciones transaccionActualizada = grupoTransaccionesService.save(transaccionExistente);
         return ResponseEntity.ok(transaccionActualizada);
     }
 
     @DeleteMapping("/{grupoId}")
-    public ResponseEntity<String> eliminarGrupo(@PathVariable Long grupoId) {
-        // Busca el grupo por su ID
+    public ResponseEntity<String> eliminarGrupo(@PathVariable Long grupoId, Authentication authentication) {
+        String emailUsuario = authentication.getName();
         Grupo grupo = grupoService.findById(grupoId);
         if (grupo == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Grupo no encontrado.");
         }
-
-        // Elimina las transacciones del grupo
+        boolean perteneceAlGrupo = grupo.getUsuarios().stream()
+            .anyMatch(usuario -> usuario.getEmail().equals(emailUsuario));
+        if (!perteneceAlGrupo) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("No tienes permiso para cerrar este grupo.");
+        }
         List<GrupoTransacciones> transacciones = grupo.getTransacciones();
         for (GrupoTransacciones transaccion : transacciones) {
             grupoTransaccionesService.delete(transaccion.getId());
         }
-
-        // Elimina las transacciones pendientes del grupo
         List<TransaccionesPendientes> transaccionesPendientes = transaccionesPendientesService.findByGrupoId(grupoId);
         for (TransaccionesPendientes transaccionPendiente : transaccionesPendientes) {
             transaccionesPendientesService.delete(transaccionPendiente.getId());
         }
-
-        // Finalmente, elimina el grupo
         grupoService.delete(grupoId);
-
         return ResponseEntity.ok("Grupo y todas sus transacciones eliminadas exitosamente.");
     }
 
     @GetMapping("/{grupoId}/usuarios")
-    public ResponseEntity<List<User>> obtenerUsuariosDelGrupo(@PathVariable Long grupoId) {
+    public ResponseEntity<List<User>> obtenerUsuariosDelGrupo(@PathVariable Long grupoId, Authentication authentication) {
+        String emailUsuario = authentication.getName();
         Grupo grupo = grupoService.findById(grupoId);
-        System.out.println("\n\n " + grupo + "\n\n " + grupoId);
         if (grupo == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Collections.emptyList());
+        }
+        boolean perteneceAlGrupo = grupo.getUsuarios().stream()
+        .anyMatch(usuario -> usuario.getEmail().equals(emailUsuario));
+        if (!perteneceAlGrupo) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Collections.emptyList());
         }
         List<User> usuarios = grupo.getUsuarios();
         return ResponseEntity.ok(usuarios);
@@ -307,7 +346,7 @@ public class GrupoController {
         for (String email : miembrosEmails) {
             User usuario = userService.findByEmail(email);
             LocalDate fechaHoy = LocalDate.now();
-            TransaccionesPendientes transaccionPendiente = new TransaccionesPendientes(0.0, usuario, grupo.getNombre(), "Grupo", fechaHoy);
+            TransaccionesPendientes transaccionPendiente = new TransaccionesPendientes(0.0, usuario, grupo.getNombre(), "Grupo", fechaHoy,null,null);
             // Establecer el correo del creador como sentByEmail
             transaccionPendiente.setSentByEmail(creadorEmail);
             transaccionPendiente.setGrupoId(grupo.getId());
@@ -320,8 +359,9 @@ public class GrupoController {
     @GetMapping("/{grupoId}/verificar-usuario")
     public ResponseEntity<String> verificarUsuarioEnGrupoOInvitado(
             @PathVariable Long grupoId,
-            @RequestParam String email) {
-        // Busca el grupo por su ID
+            @RequestParam String email,
+            Authentication authentication) {
+        String emailUsuario = authentication.getName();
         Grupo grupo = grupoService.findById(grupoId);
         if (grupo == null) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Grupo no encontrado.");
@@ -345,7 +385,6 @@ public class GrupoController {
                 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("El usuario ya tiene una invitación pendiente para el grupo.");
             }
         }
-
         return ResponseEntity.ok("El usuario no está en el grupo ni tiene una invitación pendiente.");
     }
